rename package name refs #6
