+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
     "cSpell.words": [
-        "linkwatch",
+        "linkstat",
         "pythonpath",
         "testpaths"
     ]
