+++ README_jp.md
@@ -1,2 +1,2 @@
-# LinkWatch
+# linkstat
