+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "linkwatch"
+name = "linkstat"
 authors = [{ name = "DogFortune", email = "dogfortune.2000@gmail.com" }]
 description = "Perform connectivity checks on URLs listed in the Markdown"
 license-files = ["LICENSE"]
@@ -22,16 +22,16 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project.scripts]
-linkwatch = "linkwatch.app:main"
+linkstat = "linkstat.app:main"
 
 [project.urls]
-Repository = "https://github.com/DogFortune/LinkWatch"
-Issues = "https://github.com/DogFortune/LinkWatch/issues"
-Changelog = "https://github.com/DogFortune/LinkWatch/releases"
+Repository = "https://github.com/DogFortune/linkstat"
+Issues = "https://github.com/DogFortune/linkstat/issues"
+Changelog = "https://github.com/DogFortune/linkstat/releases"
 
 [tool.pytest.ini_options]
 pythonpath = "src"
 testpaths = ["tests"]
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/linkwatch"]
+packages = ["src/linkstat"]
