+++ src/linkstat/__main__.py
