+++ src/linkstat/analyzer.py
@@ -1,8 +1,8 @@
 from pathlib import Path
 from urllib.request import urlopen
 from urllib.error import HTTPError, URLError
-from linkwatch.enums import Result
-from linkwatch.reporter import ReportData
+from linkstat.enums import Result
+from linkstat.reporter import ReportData
 from dataclasses import dataclass
 import re
 from tqdm import tqdm
