+++ src/linkstat/enums.py
