+++ tests/test_reporter.py
@@ -1,6 +1,6 @@
 import pytest
-from linkwatch import reporter
-from linkwatch import analyzer
+from linkstat import reporter
+from linkstat import analyzer
 from tempfile import TemporaryDirectory
 from pathlib import Path
 import os
